backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain-openai
langchain-core
pydantic
pydantic-settings
PyPDF2
python-multipart
supabase
sentence-transformers
numpy
